requirements.txt
customtkinter 
Pillow